Migrating HasUserResizedLvHeader to child
